MOV: Moved Static Endpoints to Blueprints
- New module for static endpoints as part of a registerable blueprint.
